Keep the FFT size guard within a 32-bit usize [skip ci]
`n > (1 << 40)` overflows usize on wasm32, a hard error there, so the
Python WebAssembly variant stopped compiling. The comparison is made in
u64.

+++ crates/zipp-vm/src/vm/py_tensor.rs
@@ -1316,7 +1316,7 @@ impl Vm<'_> {
         let mode = u32::try_from(int_arg(a, 5)?).ok()?;
         let inverse = num_arg(a, 6)? != 0.0;
         let scale = num_arg(a, 7)?;
-        if mode > 3 || n == 0 || n > (1 << 40) {
+        if mode > 3 || n == 0 || n as u64 > (1u64 << 40) {
             return None;
         }
         let half = (n >> 1) + 1;
